+++ spec/kommando/in_spec.rb
@@ -17,7 +17,10 @@ describe Kommando do
       it 'takes input in for a command' do
         tmpfile = Tempfile.new "test"
         k = Kommando.new "nano #{tmpfile.path}", {
-          timeout: 0.5
+          timeout: 0.5,
+          env: {
+            TERM: "vt100"  #NOTE: fixes running in docker build
+          }
         }
         k.in << "\x1B\x1Bx"
         k.run