+++ demos/ping_monitor.rb
@@ -8,4 +8,11 @@ k.out.every /time=(\d+\.\d+)\s/ do |m|
   puts "x" * (time*300.to_i)
 end
 
+k.out.on(/^PING.+\n$/).every(/^(.+)\r\n/) do |m|
+  unless m[1].start_with? "64 bytes from"
+    puts "ERR: unexpected reply: #{m[1]}"
+    exit 1
+  end
+end
+
 k.run